feat: optional bypass value for default env
This feature is an optional bypass for the existing default env records provided by workload helper.

The existing behaviour of env is unchanged, only providing the  disableDefaultEnv value with a list of names will cause any of the default env to be removed.

this has been tested on a application where I needed to override the HOSTNAME default with a custom record.